refactor(retry-queue): move `addRetryQueueListener` to separate file
